+++ src/Microsoft.Diagnostics.Runtime/AbstractDac/IAbstractModuleProvider.cs
@@ -8,6 +8,7 @@ namespace Microsoft.Diagnostics.Runtime.AbstractDac
 {
     internal interface IAbstractModuleProvider
     {
+        ClrModuleInfo GetModuleInfo(ulong module);
         MetadataImport? GetMetadataImport(ulong module);
         IEnumerable<(ulong MethodTable, int Token)> EnumerateTypeDefMap(ulong module);
         IEnumerable<(ulong MethodTable, int Token)> EnumerateTypeRefMap(ulong module);
